requirements.txt
fastapi==0.95.0
uvicorn==0.20.0
torch
torchvision
transformers==4.30.0
pillow==9.4.0